+++ socialhome/frontend/components/streams/Stream.vue
@@ -43,6 +43,7 @@ import "frontend/components/streams/StreamElement.vue"
 import PublicStampedElement from "frontend/components/streams/stamped_elements/PublicStampedElement.vue"
 import FollowedStampedElement from "frontend/components/streams/stamped_elements/FollowedStampedElement.vue"
 import LimitedStampedElement from "frontend/components/streams/stamped_elements/LimitedStampedElement.vue"
+import LocalStampedElement from "frontend/components/streams/stamped_elements/LocalStampedElement.vue"
 import TagStampedElement from "frontend/components/streams/stamped_elements/TagStampedElement.vue"
 import ProfileStampedElement from "frontend/components/streams/stamped_elements/ProfileStampedElement.vue"
 import "frontend/components/streams/LoadingElement.vue"
@@ -60,10 +61,11 @@ export default Vue.component("stream", {
     },
     components: {
         FollowedStampedElement,
-        PublicStampedElement,
-        ProfileStampedElement,
-        TagStampedElement,
         LimitedStampedElement,
+        LocalStampedElement,
+        ProfileStampedElement,
+        PublicStampedElement,
+        TagStampedElement,
     },
     data() {
         return {
@@ -91,6 +93,8 @@ export default Vue.component("stream", {
                     return "FollowedStampedElement"
                 case "limited":
                     return "LimitedStampedElement"
+                case "local":
+                    return "LocalStampedElement"
                 case "public":
                     return "PublicStampedElement"
                 case "tag":
@@ -132,6 +136,9 @@ export default Vue.component("stream", {
                     break
                 case "limited":
                     this.$store.dispatch(streamStoreOperations.getLimitedStream, options)
+                    break
+                case "local":
+                    this.$store.dispatch(streamStoreOperations.getLocalStream, options)
                     break
                 case "public":
                     this.$store.dispatch(streamStoreOperations.getPublicStream, options)